docker_all/components/remote_real_x3plus/app/requirements.txt
pyzmq
pyserial
